requirements3.txt
GitPython>=3.1.1
termcolor>=1.1.0
six>=1.14.0
pytest>=6.2.2
isort>=5.9.3
black>=23.7.1
